Refactor: Generalize column auto-resize for FilesBrowser, introduce ContentAwareTreeView (#3874) (#360)
