+++ HISTORY.rst
@@ -16,6 +16,7 @@ History
 - Prevent locking of layers based on gpkg metadata (nens/rana#3942)
 - Replace layers that are already opened instead of opening duplicates (nens/#rana#3890)
 - Fix returning to FilesBrowser after delete via FileView
+- Improve FilesBrowser column resizing for long filenames (nens/rana#3874)
 
 
 1.2.9 (2026-03-30)
